# feat: optimize encode and decode on M3 Mac