--- a/src/writer.rs
+++ b/src/writer.rs
@@ -139,13 +139,13 @@ impl Encoder {
 
         // Fill the buffer with the 2-bit representation of the nucleotides
         self.clear();
-        if bitnuc::encode(primary, &mut self.sbuffer).is_err() {
+        if crate::nuc::encode(primary, &mut self.sbuffer).is_err() {
             self.clear();
             if self
                 .policy
                 .handle(primary, &mut self.s_ibuf, &mut self.rng)?
             {
-                bitnuc::encode(&self.s_ibuf, &mut self.sbuffer)?;
+                crate::nuc::encode(&self.s_ibuf, &mut self.sbuffer)?;
             } else {
                 return Ok(None);
             }
@@ -178,8 +178,8 @@ impl Encoder {
         }
 
         self.clear();
-        if bitnuc::encode(primary, &mut self.sbuffer).is_err()
-            || bitnuc::encode(extended, &mut self.xbuffer).is_err()
+        if crate::nuc::encode(primary, &mut self.sbuffer).is_err()
+            || crate::nuc::encode(extended, &mut self.xbuffer).is_err()
         {
             self.clear();
             if self
@@ -189,8 +189,8 @@ impl Encoder {
                     .policy
                     .handle(extended, &mut self.x_ibuf, &mut self.rng)?
             {
-                bitnuc::encode(&self.s_ibuf, &mut self.sbuffer)?;
-                bitnuc::encode(&self.x_ibuf, &mut self.xbuffer)?;
+                crate::nuc::encode(&self.s_ibuf, &mut self.sbuffer)?;
+                crate::nuc::encode(&self.x_ibuf, &mut self.xbuffer)?;
             } else {
                 return Ok(None);
             }