Hello this is a very cool project! 

This PR adds an optimized implementation for `aarch64` machines via `NEON` intrininsics

### Bench

The changes include a very simple benchmark that can be run like

```bash
cargo run --release --example bench_simd
```

```txt
Architecture: aarch64 (ARM64)
SIMD support detected: true

Encode Benchmark Results
----------------------------------------------------------------------------------------------------------
| Length     | Iterations | Original (ms)   | SIMD (ms)       | Orig (Mnuc/s) | SIMD (Mnuc/s) | Speedup  |
----------------------------------------------------------------------------------------------------------
| 32         | 1000       | 0               | 0               | 417.62        | 405.27        | 0.97x    |
| 64         | 1000       | 0               | 0               | 391.64        | 409.93        | 1.05x    |
| 100        | 1000       | 0               | 0               | 404.45        | 486.91        | 1.20x    |
| 1000       | 1000       | 2               | 1               | 492.97        | 537.60        | 1.09x    |
| 10000      | 1000       | 7               | 1               | 1290.68       | 9450.30       | 7.32x    |
| 100000     | 1000       | 56              | 10              | 1756.59       | 9832.60       | 5.60x    |
| 1000000    | 1000       | 574             | 95              | 1741.44       | 10440.67      | 6.00x    |
| 5000000    | 1000       | 2866            | 481             | 1744.01       | 10389.16      | 5.96x    |
----------------------------------------------------------------------------------------------------------

Decode Benchmark Results
----------------------------------------------------------------------------------------------------------
| Length     | Iterations | Original (ms)   | SIMD (ms)       | Orig (Mnuc/s) | SIMD (Mnuc/s) | Speedup  |
----------------------------------------------------------------------------------------------------------
| 32         | 1000       | 0               | 0               | 4000.00       | 3898.64       | 0.97x    |
| 64         | 1000       | 0               | 0               | 5505.38       | 5052.90       | 0.92x    |
| 100        | 1000       | 0               | 0               | 5491.79       | 5217.57       | 0.95x    |
| 1000       | 1000       | 0               | 0               | 6659.25       | 6618.88       | 0.99x    |
| 10000      | 1000       | 1               | 1               | 6693.82       | 6664.63       | 1.00x    |
| 100000     | 1000       | 14              | 4               | 6712.22       | 20469.61      | 3.05x    |
| 1000000    | 1000       | 194             | 48              | 5131.55       | 20795.86      | 4.05x    |
| 5000000    | 1000       | 970             | 237             | 5151.39       | 21073.47      | 4.09x    |
----------------------------------------------------------------------------------------------------------
```

notable these changes only show a speedup at long sequence lengths (>1000) as the overhead required outweigh any performance gains. In the changes we only apply the new function if the input `>1000` with a minor impact for the added conditional check.


### Combined with parallel processing

As noted the impact is seen at larger sequence lengths so if the `parallel_processing.rs` file is updated to use larger sequence lengths the performance gains are pronounced.

```rust
let r1_size = 1_500_000;
let r2_size = 3_000_000;
let num_seq = 1000;
```

then running the example

```bash
cargo run --release --example parallel_processing
#     Finished `release` profile [optimized] target(s) in 0.03s
#      Running `target/release/examples/parallel_processing`
# Finished writing 1000 records to path: ./data/test.bq
# Elapsed time (write_single): 5.790401542s
# Finished writing 1000 records to path: ./data/test_paired.bq
# Elapsed time (write_paired): 17.304640042s
# Elapsed time (single - mmap_parallel_processing (1)): 103.262875ms
# Elapsed time (single - mmap_parallel_processing (2)): 51.064292ms
# Elapsed time (single - mmap_parallel_processing (4)): 27.45525ms
# Elapsed time (single - mmap_parallel_processing (6)): 18.840167ms
# Elapsed time (single - mmap_parallel_processing (8)): 15.057875ms
# Elapsed time (single - mmap_parallel_processing (10)): 15.177208ms
# Elapsed time (single - mmap_parallel_processing (12)): 15.125666ms
# Elapsed time (single - mmap_parallel_processing (14)): 15.347084ms
# Elapsed time (single - mmap_parallel_processing (16)): 15.18525ms
# Elapsed time (paired - mmap_parallel_processing (1)): 293.514375ms
# Elapsed time (paired - mmap_parallel_processing (2)): 141.172ms
# Elapsed time (paired - mmap_parallel_processing (4)): 77.872583ms
# Elapsed time (paired - mmap_parallel_processing (6)): 57.239875ms
# Elapsed time (paired - mmap_parallel_processing (8)): 53.115916ms
# Elapsed time (paired - mmap_parallel_processing (10)): 50.590417ms
# Elapsed time (paired - mmap_parallel_processing (12)): 48.925666ms
# Elapsed time (paired - mmap_parallel_processing (14)): 50.635833ms
# Elapsed time (paired - mmap_parallel_processing (16)): 50.277625ms
```

and with the new changes disabled

```bash
DISABLE_SIMD=1 cargo run --release --example parallel_processing
#     Finished `release` profile [optimized] target(s) in 0.03s
#      Running `target/release/examples/parallel_processing_long_bp`
# Finished writing 1000 records to path: ./data/test.bq
# Elapsed time (write_single): 6.517152875s
# Finished writing 1000 records to path: ./data/test_paired.bq
# Elapsed time (write_paired): 19.553436125s
# Elapsed time (single - mmap_parallel_processing (1)): 320.603167ms
# Elapsed time (single - mmap_parallel_processing (2)): 167.1955ms
# Elapsed time (single - mmap_parallel_processing (4)): 83.050459ms
# Elapsed time (single - mmap_parallel_processing (6)): 57.442542ms
# Elapsed time (single - mmap_parallel_processing (8)): 44.447916ms
# Elapsed time (single - mmap_parallel_processing (10)): 38.634167ms
# Elapsed time (single - mmap_parallel_processing (12)): 35.195ms
# Elapsed time (single - mmap_parallel_processing (14)): 42.031541ms
# Elapsed time (single - mmap_parallel_processing (16)): 40.494667ms
# Elapsed time (paired - mmap_parallel_processing (1)): 994.009084ms
# Elapsed time (paired - mmap_parallel_processing (2)): 500.7545ms
# Elapsed time (paired - mmap_parallel_processing (4)): 255.31825ms
# Elapsed time (paired - mmap_parallel_processing (6)): 172.418084ms
# Elapsed time (paired - mmap_parallel_processing (8)): 133.924333ms
# Elapsed time (paired - mmap_parallel_processing (10)): 110.315458ms
# Elapsed time (paired - mmap_parallel_processing (12)): 106.56625ms
# Elapsed time (paired - mmap_parallel_processing (14)): 103.104708ms
# Elapsed time (paired - mmap_parallel_processing (16)): 108.734792ms
```

This shows a sizable speedup of +2-3x in both the single thread and multi-threaded cases on the longer sequences

Thanks again! Please let me know if I should make any changes 🙏