@drbh 2025-04-18
@noamteyssier thanks for the quick response! 

That makes a lot of sense and sounds good, I'll follow up with a PR in the bitnuc crate soon!