@noamteyssier 2025-04-18
Thanks I appreciate it! I'll close this one here and I'll see you on the other side.